requirements.txt
redis==4.5.4
PyYAML==6.0
psutil==5.9.5
asyncio-mqtt==0.16.1
motor==3.7.1 